Add files via upload
In-order Traversal: In a binary tree, this method visits the left subtree first, then the root node, and finally the right subtree. For binary search trees (BST), in-order traversal retrieves data in sorted order, making it useful for operations that require sorted data.

Pre-order Traversal: This method visits the root node first, followed by the left subtree and then the right subtree. Pre-order traversal is beneficial for creating a copy of the tree or for prefix expression evaluation in expression trees.

Post-order Traversal: In this method, the left subtree is visited first, followed by the right subtree, and finally the root node. Post-order traversal is often used for deleting a tree or evaluating postfix expressions.

Level-order Traversal: This traversal visits nodes level by level, starting from the root and moving down to each subsequent level. It is typically implemented using a queue and is useful for breadth-first exploration of the tree.